Extract OpenCallApplication withdrawal logic from list & details pages
into a reusable WithdrawApplicationModal component
