+++ frontend/containers/dashboard/open-call-applications/withdraw-application-modal/index.ts
@@ -1,0 +1,2 @@
+export { default } from './component';
+export type { WithdrawApplicationModalProps } from './types';
